+++ driver/src/main/java/oracle/nosql/driver/package-info.java
@@ -89,5 +89,54 @@
  * it's possible to have too many threads and that more threads does not
  * equal more performance. The optimal number depends on request latency and
  * other I/O performed. It is best to experiment.
+ * <p>
+ * <strong>Logging in the SDK</strong>
+ * <p>
+ * The SDK uses logging as provided by the <i>java.util.logging</i> package.
+ * By default the SDK will log to the console at level INFO. If nothing
+ * does wrong there is little or no logging. There are 2 simple ways to
+ * configure logging if additional information is desired.
+ * <ol>
+ * <li>Create and use a logging configuration file and pass it to the
+ * application using a system property. E.g.
+ * <pre>
+ *     $ java -Djava.util.logging.config.file=path.to.logging.properties ...
+ *  or if using mvn:exec
+ *     $ mvn exec:java -Djava.util.logging.config.file=path.to.logging.properties ...
+ * </pre>
+ * The content of the properties file is as documented by
+ * <i>java.util.logging</i>. See below for an example. This mechanism makes
+ * sense if the application doesn't have its own logging needs.
+ * <li>
+ * Creating an instance of {@link java.util.logging.Logger} in the application
+ * and passing it to the SDK using {@link NoSQLHandleConfig#setLogger}. This
+ * option can be used if the application wants to create its own custom
+ * logging configuration. It can also use a logging configuration file.
+ * </li>
+ * </ol>
+ * Here is an example of a logging properties file that will log
+ * both the SDK and the underlying Netty networking at level FINE. It will
+ * log to both the console and a local file called "driver.log."
+ * <pre>
+handlers=java.util.logging.FileHandler, java.util.logging.ConsoleHandler
+
+# File config
+java.util.logging.FileHandler.level=ALL
+java.util.logging.FileHandler.formatter=java.util.logging.SimpleFormatter
+java.util.logging.FileHandler.pattern=driver.log
+java.util.logging.FileHandler.count=1
+java.util.logging.FileHandler.limit=50000
+
+# Console config
+java.util.logging.ConsoleHandler.level=ALL
+java.util.logging.ConsoleHandler.formatter=java.util.logging.SimpleFormatter
+
+# Use a non-default, single-line, simple format pattern
+java.util.logging.SimpleFormatter.format=%1$tF %1$tT %4$-7s %5$s %n
+
+# Level can be SEVERE, WARNING, INFO, FINE, ALL, OFF
+oracle.nosql.level=FINE
+io.netty.level=FINE
+ * </pre>
  */
 package oracle.nosql.driver;